feat: add useCase for updating a board
