+++ backend/src/libs/test-utils/mocks/factories/dto/updateBoardDto-factory.mock.ts
@@ -3,10 +3,16 @@ import { BoardUserFactory } from '../boardUser-factory.mock';
 import { buildTestFactory } from '../generic-factory.mock';
 import { BoardDtoFactory } from './boardDto-factory.mock';
 
+const boardData = BoardDtoFactory.create();
+
 const mockUpdateBoardDto = () => {
 	return {
 		responsible: BoardUserFactory.create(),
-		...BoardDtoFactory.create()
+		boardId: boardData._id,
+		completionHandler: () => {
+			return;
+		},
+		...boardData
 	};
 };
