+++ backend/src/modules/boards/applications/update.board.application.ts
@@ -1,9 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { UpdateBoardDto } from '../dto/update-board.dto';
 import { UpdateBoardApplicationInterface } from '../interfaces/applications/update.board.application.interface';
 import { UpdateBoardServiceInterface } from '../interfaces/services/update.board.service.interface';
 import { TYPES } from '../interfaces/types';
-import UpdateBoardUserDto from '../../boardUsers/dto/update-board-user.dto';
 import { BoardPhaseDto } from 'src/libs/dto/board-phase.dto';
 
 @Injectable()
@@ -13,22 +11,8 @@ export class UpdateBoardApplication implements UpdateBoardApplicationInterface {
 		private updateBoardService: UpdateBoardServiceInterface
 	) {}
 
-	update(boardId: string, boardData: UpdateBoardDto) {
-		return this.updateBoardService.update(boardId, boardData);
-	}
-
 	mergeBoards(subBoardId: string, userId: string, socketId?: string) {
 		return this.updateBoardService.mergeBoards(subBoardId, userId, socketId);
-	}
-
-	updateBoardParticipants(boardData: UpdateBoardUserDto) {
-		const { addBoardUsers, removeBoardUsers, boardUserToUpdateRole } = boardData;
-
-		if (boardUserToUpdateRole) {
-			return this.updateBoardService.updateBoardParticipantsRole(boardUserToUpdateRole);
-		}
-
-		return this.updateBoardService.updateBoardParticipants(addBoardUsers, removeBoardUsers);
 	}
 
 	updatePhase(payload: BoardPhaseDto) {
