+++ backend/src/modules/boards/services/update.board.service.ts
@@ -1,24 +1,13 @@
 import Team from 'src/modules/teams/entities/team.schema';
-import { UpdateBoardUserServiceInterface } from '../../boardUsers/interfaces/services/update.board.user.service.interface';
-import BoardUserDto from 'src/modules/boardUsers/dto/board.user.dto';
 import { Inject, Injectable, Logger } from '@nestjs/common';
-import { getIdFromObjectId } from 'src/libs/utils/getIdFromObjectId';
-import isEmpty from 'src/libs/utils/isEmpty';
 import { TeamDto } from 'src/modules/communication/dto/team.dto';
 import { CommunicationServiceInterface } from 'src/modules/communication/interfaces/slack-communication.service.interface';
 import * as CommunicationsType from 'src/modules/communication/interfaces/types';
-import * as Votes from 'src/modules/votes/interfaces/types';
 import * as Boards from 'src/modules/boards/interfaces/types';
 import * as BoardUsers from 'src/modules/boardUsers/interfaces/types';
-import User from 'src/modules/users/entities/user.schema';
-import { UpdateBoardDto } from '../dto/update-board.dto';
-import { ResponsibleType } from '../interfaces/responsible.interface';
 import { UpdateBoardServiceInterface } from '../interfaces/services/update.board.service.interface';
 import Board from '../entities/board.schema';
-import BoardUser from '../../boardUsers/entities/board.user.schema';
 import SocketGateway from 'src/modules/socket/gateway/socket.gateway';
-import Column from '../../columns/entities/column.schema';
-import ColumnDto from '../../columns/dto/column.dto';
 import { BoardRepositoryInterface } from '../repositories/board.repository.interface';
 import { BOARD_PHASE_SERVER_UPDATED } from 'src/libs/constants/phase';
 import { FRONTEND_URL } from 'src/libs/constants/frontend';
@@ -30,14 +19,11 @@ import { SlackMessageDto } from 'src/modules/communication/dto/slack.message.dto
 import { SLACK_ENABLE, SLACK_MASTER_CHANNEL_ID } from 'src/libs/constants/slack';
 import { ConfigService } from '@nestjs/config';
 import { BoardPhases } from 'src/libs/enum/board.phases';
-import { GetBoardUserServiceInterface } from 'src/modules/boardUsers/interfaces/services/get.board.user.service.interface';
-import { CreateBoardUserServiceInterface } from 'src/modules/boardUsers/interfaces/services/create.board.user.service.interface';
 import { DeleteBoardUserServiceInterface } from 'src/modules/boardUsers/interfaces/services/delete.board.user.service.interface';
 import { generateNewSubColumns } from '../utils/generate-subcolumns';
 import { mergeCardsFromSubBoardColumnsIntoMainBoard } from '../utils/merge-cards-from-subboard';
 import { UpdateFailedException } from 'src/libs/exceptions/updateFailedBadRequestException';
 import { BoardNotFoundException } from 'src/libs/exceptions/boardNotFoundException';
-import { DeleteVoteServiceInterface } from 'src/modules/votes/interfaces/services/delete.vote.service.interface';
 
 @Injectable()
 export default class UpdateBoardService implements UpdateBoardServiceInterface {
@@ -49,14 +35,6 @@ export default class UpdateBoardService implements UpdateBoardServiceInterface {
 		@Inject(CommunicationsType.TYPES.services.SlackSendMessageService)
 		private slackSendMessageService: SendMessageServiceInterface,
 		private socketService: SocketGateway,
-		@Inject(Votes.TYPES.services.DeleteVoteService)
-		private deleteVoteService: DeleteVoteServiceInterface,
-		@Inject(BoardUsers.TYPES.services.CreateBoardUserService)
-		private readonly createBoardUserService: CreateBoardUserServiceInterface,
-		@Inject(BoardUsers.TYPES.services.GetBoardUserService)
-		private readonly getBoardUserService: GetBoardUserServiceInterface,
-		@Inject(BoardUsers.TYPES.services.UpdateBoardUserService)
-		private readonly updateBoardUserService: UpdateBoardUserServiceInterface,
 		@Inject(BoardUsers.TYPES.services.DeleteBoardUserService)
 		private readonly deleteBoardUserService: DeleteBoardUserServiceInterface,
 		@Inject(Boards.TYPES.repositories.BoardRepository)
@@ -64,106 +42,6 @@ export default class UpdateBoardService implements UpdateBoardServiceInterface {
 		private eventEmitter: EventEmitter2,
 		private configService: ConfigService
 	) {}
-
-	async update(boardId: string, boardData: UpdateBoardDto) {
-		/**
-		 * Only can change the maxVotes if:
-		 * - new maxVotes not empty
-		 * - current highest votes equals to zero
-		 * - or current highest votes lower than new maxVotes
-		 */
-
-		if (!isEmpty(boardData.maxVotes)) {
-			const highestVotes = await this.getHighestVotesOnBoard(boardId);
-
-			if (highestVotes > Number(boardData.maxVotes)) {
-				throw new UpdateFailedException(
-					`You can't set a lower value to max votes. Please insert a value higher or equals than ${highestVotes}!`
-				);
-			}
-		}
-
-		const board = await this.boardRepository.getBoard(boardId);
-
-		if (!board) {
-			throw new BoardNotFoundException();
-		}
-
-		// Destructuring board/boardData variables
-		const { isSubBoard } = board;
-		const { responsible } = boardData;
-
-		const currentResponsible = await this.getBoardResponsibleInfo(boardId);
-		const newResponsible: ResponsibleType = {
-			id: (responsible?.user as User)._id,
-			email: (responsible?.user as User).email
-		};
-
-		/**
-		 * Validate if:
-		 * - have users on request
-		 * - and the current responsible isn't the new responsible
-		 */
-		if (boardData.users && String(currentResponsible?.id) !== String(newResponsible?.id)) {
-			this.changeResponsibleOnBoard(
-				isSubBoard,
-				boardId,
-				boardData.mainBoardId,
-				boardData.users,
-				String(currentResponsible.id),
-				String(newResponsible.id)
-			);
-		}
-
-		/**
-		 * Updates the board's settings fields
-		 *
-		 * */
-
-		board.title = boardData.title;
-		board.maxVotes = boardData.maxVotes;
-		board.hideCards = boardData.hideCards;
-		board.addCards = boardData.addCards;
-		board.hideVotes = boardData.hideVotes;
-		board.postAnonymously = boardData.postAnonymously;
-		board.isPublic = boardData.isPublic;
-
-		/**
-		 * If the board is a regular, then updates its columns
-		 *
-		 * */
-		if (!isSubBoard && isEmpty(board.dividedBoards)) {
-			board.columns = await this.updateRegularBoard(boardId, boardData, board);
-		}
-
-		const updatedBoard = await this.boardRepository.updateBoard(boardId, board, true);
-
-		if (!updatedBoard) throw new UpdateFailedException();
-
-		if (boardData.socketId) {
-			this.socketService.sendUpdatedBoard(boardId, boardData.socketId);
-		}
-
-		if (
-			updatedBoard &&
-			newResponsible &&
-			currentResponsible &&
-			String(currentResponsible?.id) !== String(newResponsible?.id) &&
-			board.slackChannelId &&
-			updatedBoard.slackEnable &&
-			updatedBoard.isSubBoard
-		) {
-			this.handleResponsibleSlackMessage(
-				newResponsible,
-				currentResponsible,
-				board._id,
-				board.slackChannelId,
-				board.boardNumber
-			);
-		}
-
-		return updatedBoard;
-	}
 
 	async mergeBoards(subBoardId: string, userId: string, socketId?: string) {
 		const [subBoard, board] = await Promise.all([
@@ -176,7 +54,7 @@ export default class UpdateBoardService implements UpdateBoardServiceInterface {
 		}
 
 		const columnsWithMergedCards = this.getColumnsFromMainBoardWithMergedCards(subBoard, board);
-		let mergedBoard;
+		let mergedBoard: Board;
 
 		await this.boardRepository.startTransaction();
 
@@ -240,36 +118,36 @@ export default class UpdateBoardService implements UpdateBoardServiceInterface {
 		);
 	}
 
-	async updateBoardParticipants(addUsers: BoardUserDto[], removeUsers: string[]) {
-		try {
-			let createdBoardUsers: BoardUser[] = [];
+	// async updateBoardParticipants(addUsers: BoardUserDto[], removeUsers: string[]) {
+	// 	try {
+	// 		let createdBoardUsers: BoardUser[] = [];
 
-			if (addUsers.length > 0)
-				createdBoardUsers = await this.createBoardUserService.saveBoardUsers(addUsers);
+	// 		if (addUsers.length > 0)
+	// 			createdBoardUsers = await this.createBoardUserService.saveBoardUsers(addUsers);
 
-			if (removeUsers.length > 0) await this.deleteBoardUsers(removeUsers);
+	// 		if (removeUsers.length > 0) await this.deleteBoardUsers(removeUsers);
 
-			return createdBoardUsers;
-		} catch (error) {
-			throw new UpdateFailedException();
-		}
-	}
+	// 		return createdBoardUsers;
+	// 	} catch (error) {
+	// 		throw new UpdateFailedException();
+	// 	}
+	// }
 
-	async updateBoardParticipantsRole(boardUserToUpdateRole: BoardUserDto) {
-		const user = boardUserToUpdateRole.user as User;
+	// async updateBoardParticipantsRole(boardUserToUpdateRole: BoardUserDto) {
+	// 	const user = boardUserToUpdateRole.user as User;
 
-		const updatedBoardUser = await this.updateBoardUserService.updateBoardUserRole(
-			boardUserToUpdateRole.board,
-			user._id,
-			boardUserToUpdateRole.role
-		);
+	// 	const updatedBoardUser = await this.updateBoardUserService.updateBoardUserRole(
+	// 		boardUserToUpdateRole.board,
+	// 		user._id,
+	// 		boardUserToUpdateRole.role
+	// 	);
 
-		if (!updatedBoardUser) {
-			throw new UpdateFailedException();
-		}
+	// 	if (!updatedBoardUser) {
+	// 		throw new UpdateFailedException();
+	// 	}
 
-		return updatedBoardUser;
-	}
+	// 	return updatedBoardUser;
+	// }
 
 	async updatePhase(boardPhaseDto: BoardPhaseDto) {
 		try {
@@ -298,153 +176,6 @@ export default class UpdateBoardService implements UpdateBoardServiceInterface {
 	}
 
 	/* --------------- HELPERS --------------- */
-
-	/**
-	 * Method to get the highest value of votesCount on Board Users
-	 * @param boardId String
-	 * @return number
-	 */
-	private async getHighestVotesOnBoard(boardId: string): Promise<number> {
-		const votesCount = await this.getBoardUserService.getVotesCount(boardId);
-
-		return votesCount.reduce(
-			(prev, current) => (current.votesCount > prev ? current.votesCount : prev),
-			0
-		);
-	}
-
-	/**
-	 * Method to get current responsible to a specific board
-	 * @param boardId String
-	 * @return Board User
-	 * @private
-	 */
-	private async getBoardResponsibleInfo(boardId: string): Promise<ResponsibleType | undefined> {
-		const boardUser = await this.getBoardUserService.getBoardResponsible(boardId);
-
-		if (!boardUser) {
-			return undefined;
-		}
-
-		const user = boardUser?.user as User;
-
-		return { id: user._id, email: user.email };
-	}
-
-	/**
-	 * Method to update all boardUsers role
-	 * @param boardId String
-	 * @param boardUsers BoardUserDto[]
-	 * @param currentResponsibleId String
-	 * @param newResponsibleId String
-	 * @return void
-	 * @private
-	 */
-	private async updateBoardUsersRole(
-		boardId: string,
-		boardUsers: BoardUserDto[],
-		currentResponsibleId: string,
-		newResponsibleId: string
-	) {
-		const promises = boardUsers
-			.filter((boardUser) =>
-				[getIdFromObjectId(currentResponsibleId), newResponsibleId].includes(
-					(boardUser.user as User)._id
-				)
-			)
-			.map((boardUser) => {
-				const typedBoardUser = boardUser.user as User;
-
-				return this.updateBoardUserService.updateBoardUserRole(
-					boardId,
-					typedBoardUser._id,
-					boardUser.role
-				);
-			});
-		await Promise.all(promises);
-	}
-
-	/**
-	 * Method to change board responsible
-	 * @return void
-	 */
-	private changeResponsibleOnBoard(
-		isSubBoard: boolean,
-		boardId: string,
-		mainBoardId: string,
-		users: BoardUserDto[],
-		currentResponsibleId: string,
-		newResponsibleId: string
-	) {
-		if (isSubBoard) {
-			this.updateBoardUsersRole(boardId, users, currentResponsibleId, newResponsibleId);
-		}
-
-		this.updateBoardUsersRole(mainBoardId, users, currentResponsibleId, newResponsibleId);
-	}
-
-	private async updateRegularBoard(boardId: string, boardData: UpdateBoardDto, board: Board) {
-		/**
-		 * Validate if:
-		 * - have columns to delete
-		 * Returns the votes to the user
-		 */
-		if (boardData.deletedColumns && !isEmpty(boardData.deletedColumns)) {
-			const cardsToDelete = boardData.deletedColumns.flatMap((deletedColumnId: string) => {
-				return board.columns.find((column) => column._id.toString() === deletedColumnId)?.cards;
-			});
-
-			await this.deleteVoteService.deleteCardVotesFromColumn(boardId, cardsToDelete);
-		}
-
-		/**
-		 * Updates the columns
-		 *
-		 * */
-		const columns = boardData.columns.flatMap((col: Column | ColumnDto) => {
-			if (col._id) {
-				const columnBoard = board.columns.find(
-					(colBoard) => colBoard._id.toString() === col._id.toString()
-				);
-
-				if (boardData.deletedColumns) {
-					const columnToDelete = boardData.deletedColumns.some(
-						(colId) => colId === col._id.toString()
-					);
-
-					if (columnToDelete) {
-						return [];
-					}
-				}
-
-				if (columnBoard) {
-					return [{ ...columnBoard, title: col.title }];
-				}
-			}
-
-			return [{ ...col }];
-		}) as Column[];
-
-		return columns;
-	}
-
-	private async handleResponsibleSlackMessage(
-		newResponsible: ResponsibleType,
-		currentResponsible: ResponsibleType | undefined,
-		boardId: string,
-		slackChannelId: string,
-		boardNumber: number
-	) {
-		this.slackCommunicationService.executeResponsibleChange({
-			newResponsibleEmail: newResponsible.email,
-			previousResponsibleEmail: currentResponsible?.email ?? '',
-			subTeamChannelId: slackChannelId,
-			responsiblesChannelId: (await this.boardRepository.getResponsiblesSlackId(boardId))
-				?.slackChannelId,
-			teamNumber: boardNumber,
-			email: newResponsible.email
-		});
-	}
 
 	private checkIfIsLastBoardToMerge(dividedBoards: Board[]): boolean {
 		const count = dividedBoards.reduce((prev, currentValue) => {
@@ -503,9 +234,9 @@ export default class UpdateBoardService implements UpdateBoardServiceInterface {
 		}
 	}
 
-	private async deleteBoardUsers(boardUsers: string[]) {
-		const deletedCount = await this.deleteBoardUserService.deleteBoardUsers(boardUsers);
+	// private async deleteBoardUsers(boardUsers: string[]) {
+	// 	const deletedCount = await this.deleteBoardUserService.deleteBoardUsers(boardUsers);
 
-		if (deletedCount <= 0) throw new UpdateFailedException();
-	}
+	// 	if (deletedCount <= 0) throw new UpdateFailedException();
+	// }
 }
